chore(lottery-v2): Break two data functions into their own files
